criando exclusao de caloteiros

+++ index.php
@@ -44,7 +44,7 @@
                         <td><?php echo $usuario['divida'] ?></td>
                         <td><?php echo $usuario['cidade'] ?></td>
                         <td>
-                            <a href="excluir.php?id=<?= $usuario['id'] ?>" style="color: red; text-decoration:none">
+                            <a href="excluir.php?id=<?= $usuario['id'] ?>" style="color: red; text-decoration:none" onclick="return confirm('Tem certeza que essa pessoa não te deve mais nada?')">
                                 <svg class="bi bi-trash-fill" width="1.5em" height="1.5em" viewBox="0 0 16 16" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
                                     <path fill-rule="evenodd" d="M2.5 1a1 1 0 0 0-1 1v1a1 1 0 0 0 1 1H3v9a2 2 0 0 0 2 2h6a2 2 0 0 0 2-2V4h.5a1 1 0 0 0 1-1V2a1 1 0 0 0-1-1H10a1 1 0 0 0-1-1H7a1 1 0 0 0-1 1H2.5zm3 4a.5.5 0 0 1 .5.5v7a.5.5 0 0 1-1 0v-7a.5.5 0 0 1 .5-.5zM8 5a.5.5 0 0 1 .5.5v7a.5.5 0 0 1-1 0v-7A.5.5 0 0 1 8 5zm3 .5a.5.5 0 0 0-1 0v7a.5.5 0 0 0 1 0v-7z"/>
                                 </svg>
